moved Download, Grayscale, Save, Smooth, and SoftBin into new package "image"
