+++ pyobs/images/processors/image/__init__.py
@@ -1,0 +1,5 @@
+from .download import Download
+from .grayscale import Grayscale
+from .save import Save
+from .smooth import Smooth
+from .softbin import SoftBin
