+++ pyobs/images/processors/image/grayscale.py
