+++ pyobs/images/processors/image/smooth.py
