+++ pyobs/images/processors/misc/__init__.py
@@ -6,12 +6,7 @@ Misc
 from .addmask import AddMask
 from .calibration.calibration import Calibration
 from .createfilename import CreateFilename
-from .download import Download
-from .grayscale import Grayscale
 from .removebackground import RemoveBackground
-from .smooth import Smooth
-from .softbin import SoftBin
-from .save import Save
 from .circular_mask import CircularMask
 from .image_source_filter import ImageSourceFilter
 from .catalog_circular_mask import CatalogCircularMask
